+++ lib/redis_cluster/circuit.rb
@@ -5,32 +5,57 @@ class RedisCluster
   # Circuit is circuit breaker for RedisCluster.
   class Circuit
 
-    attr_reader :fail_count, :ban_until
+    attr_reader :fail_count, :ban_until, :causes, :trigger, :middlewares
 
-    def initialize(threshold, interval)
-      @ban_until = Time.now
+    # Create new circuit
+    #
+    # @param [int] threshold: threshold before the circuit open
+    # @param [time] interval: use to increase ban_until when the circuit open
+    # @param [RedisCluster::Middlewares] middlewares: Redis cluster middleware
+    def initialize(threshold, interval, middlewares)
+      @ban_until = Time.at 0
       @fail_count = 0
       @last_fail_time = Time.now
       @fail_threshold = threshold
       @interval_time = interval
+      @middlewares = middlewares
+
+      @causes = []
+      @trigger = nil
     end
 
     # Failed is a method to add failed count and compare it to threshold,
     # Will trip circuit if the count goes through threshold.
     #
+    # @param [Exception] err: exception that caused circuit failed
     # @return[void]
-    def failed
-      @fail_count = 0 if (@last_fail_time + @interval_time).utc < Time.now
+    def failed(err)
+      if @last_fail_time + (@interval_time * 1.5) < Time.now
+        @fail_count = 0
+        @causes = []
+      end
       @fail_count += 1
       @last_fail_time = Time.now
-      open! if @fail_count >= @fail_threshold
+      @causes << err
+      open!('failure') if @fail_count >= @fail_threshold
     end
 
     # Open! is a method to update ban time.
+    # will trigger middleware[:circuit] if exist
     #
+    # @param [string] trigger: message to indicate why the circuit open
     # @return[void]
-    def open!
-      @ban_until = (Time.now + @interval_time).utc
+    def open!(trigger)
+      @trigger = trigger
+
+      unless middlewares
+        @ban_until = Time.now + @interval_time
+        return
+      end
+
+      middlewares.invoke(:circuit, self) do
+        @ban_until = Time.now + @interval_time
+      end
     end
 
     # Open? is a method to check if the circuit breaker status.