+++ spec/redis_cluster/circuit_spec.rb
@@ -3,26 +3,26 @@ require 'redis_cluster/circuit'
 require 'pry'
 
 describe RedisCluster::Circuit do
-  subject{ described_class.new(5, 5) }
+  subject{ described_class.new(5, 5, nil) }
 
   describe '#failed' do
     it 'can up fail_count' do
-      subject.failed
+      subject.failed(nil)
       expect(subject.fail_count).to eq 1
     end
     it 'can ban after 5 failed attempt' do
-      subject.failed
-      subject.failed
-      subject.failed
-      subject.failed
-      subject.failed
+      subject.failed(nil)
+      subject.failed(nil)
+      subject.failed(nil)
+      subject.failed(nil)
+      subject.failed(nil)
   	  expect(subject.ban_until).to be > Time.now
     end
   end
-  
+
   describe '#open!' do
   	it 'can increase ban time' do
-  	  subject.open!
+  	  subject.open!("test open")
   	  expect(subject.ban_until).to be > Time.now
   	end
   end
@@ -32,7 +32,7 @@ describe RedisCluster::Circuit do
   		expect(subject.open?).to eq false
   	end
   	it 'can check if its banned' do
-  		subject.open!
+  		subject.open!("test open")
   		expect(subject.open?).to eq true
   	end
   end